Add basic support for sudo
Now when the command is "sudo", bashelp tries to show the man page for
the command that's being run by sudo.  The current way this is done is
quite simplistic: if an option is passed to sudo, the command isn't
parsed out and the man page for sudo is shown.  Still, at least it works
in what's probably the most common case, which is better than before.